+++ components/collector/src/source_collectors/robot_framework/test_cases.py
@@ -1,0 +1,7 @@
+"""Robot Framework test cases collector."""
+
+from .tests import RobotFrameworkTests
+
+
+class RobotFrameworkTestCases(RobotFrameworkTests):
+    """Collector for Robot Framework test cases."""